Refactored oah-main to include additional commands

+++ src/bash/oah-destroy.sh
@@ -3,15 +3,20 @@
 function __oah_destroy {
 	environment="$1"
 
-	# Check if the ENV is ove status
+	# TODO Check the ENV is oah status
 
+  # check if the host_mode is vagrant
+  if ['$OAH_HOST_TYPE' = 'vagrant']; then
 
-	if [ -n "$environment" ]
-	then
-	  vagrant destroy $environment
-	else
-	  echo "Unable to destroy the environment"
-	  echo "Please check the status  with oah status and  provide vagrant id  "
-	  echo "Format is "oah destroy 'id'""
+		if [ -n "$environment" ]
+		then
+			vagrant destroy $environment
+		else
+			echo "Unable to destroy the environment"
+			echo "Please check the status  with oah status and  provide vagrant id  "
+			echo "Format is "oah destroy 'id'""
+		fi
 	fi
+
+
 }
